New apkmod.sh procedure to enable decrypting + encrypting + adding broken resources

+++ tools/apkmod.sh
@@ -1,6 +1,17 @@
 #!/bin/bash
 . tools/util_sh
 
+function FindResources() {
+  local MOD_DIR=$1
+  local RES_PATTERN=$2
+  cd $MOD_DIR
+  for f in `find . -type f -name $RES_PATTERN`; do
+     n=`FixRelPath $f`
+     echo $n
+  done
+  cd - &>/dev/null
+}
+
 HELP="Usage: $0 <file.apk> <mod.dir>"
 if [ "$#" -lt "2" ]; then
   echo "$HELP"
@@ -32,61 +43,65 @@ else
   SIGN=1
 fi
 
-# Package containing compiled sources should be decompiled first
-RES=`ls $MOD_DIR/res/ | grep -v drawable`
+# Package containing compiled sources should be recompiled afterwards
+# We convert them to rel path similar to the one in the apk
+XML=`FindResources $MOD_DIR '*.xml'`
+PNG=`FindResources $MOD_DIR '*.png'`
 
 if [ -f "$APK" ]; then
-  rm -rf $OUT_DIR
-  #mkdir -p $OUT_DIR
 
-  # Decompile
-  if [ "$RES" != "" ]; then
-    ShowMessage "  [MOD|apkd] " `basename $APK` " => " `basename "$OUT_DIR"`
+  # If there are resources, we just decrypt / cp / encrypt
+  # prior to copying the images, as they may be broken (9.png...)
+
+  # $APK => $NEWAPK
+  if [ "$XML" != "" ]; then
+    # Decompile
+    ShowMessage "  [MOD|apkd] " `basename $APK`
+    rm -rf $OUT_DIR
     java -jar $APKTOOL d $APK $OUT_DIR &>> $LOG || exit 1
-  # just unzip
-  else 
-    ShowMessage "  [MOD|7zax] " `basename $APK` " => " `basename "$OUT_DIR"`
-    7za x -o"$OUT_DIR" $APK &>> $LOG || exit 1
-  fi
-
-  # Copy content
-  ShowMessage "  [MOD|cp] " `basename $MOD_DIR` " => " `basename "$OUT_DIR"`
-  cp -av $MOD_DIR/* $OUT_DIR &>> $LOG || exit 1
-
-  # Optimize png?
-	#find "$OUT_DIR/res" -name *.png | while read PNG_FILE ;
-	#do
-	#	if [ `echo "$PNG_FILE" | grep -c "\.9\.png$"` -eq 0 ] ; then
-	#		optipng -o99 "$PNG_FILE"
-	#	fi
-	#done
-
-  # Remove meta-inf on regular apk's
-  if [ "$RMMETA" = "1" ]; then
-    ShowMessage "  [MOD|rm] META-INF"
-    rm -rf $OUT_DIR/META-INF/ &>> $LOG || exit 1
-  fi
-
-  # Recompile
-  if [ "$RES" != "" ]; then
-    ShowMessage "  [MOD|apkb] " `basename $MOD_DIR` " => " `basename "$OUT_DIR"`
+    # Copy resources
+    ShowMessage "  [MOD|cp] " `basename $MOD_DIR`
+    for i in $XML; do
+       cp -av $MOD_DIR/$i $OUT_DIR &>> $LOG
+    done
+    # Recompile
+    ShowMessage "  [MOD|apkb] " `basename $MOD_DIR`
     java -jar $APKTOOL b $OUT_DIR $NEWAPK &>> $LOG || exit 1
-  # just rezip
+    mv $OUT_DIR $OUT_DIR.decompile # preserve temp dir
   else
-    ShowMessage "  [MOD|7zaa] " `basename $MOD_DIR` " => " `basename "$OUT_DIR"`
-    7za a -tzip $NEWAPK $OUT_DIR/* -mx9 &>> $LOG || exit 1
+    ShowMessage "  [MOD|cp] " `basename $APK`
+    cp -p $APK $NEWAPK
   fi
 
-  # Sign
+  # Now we 7za u (update) the images
+  if [ "$PNG" != "" ]; then
+    ShowMessage "  [MOD|7zau] " `basename $APK`
+
+    # Optimize png?
+ 		#find "$OUT_DIR/res" -name *.png | while read PNG_FILE ;
+ 		#do
+ 		#	if [ `echo "$PNG_FILE" | grep -c "\.9\.png$"` -eq 0 ] ; then
+ 		#		optipng -o99 "$PNG_FILE"
+ 		#	fi
+ 		#done
+
+    cd "$MOD_DIR"
+    set -x
+    7za -tzip u $NEWAPK $PNG &>> $LOG || exit 1
+    set +x
+    cd - >> $LOG 2>&1
+
+  fi
+
+  # Sign $NEWAPK => $APK
   if [ "$SIGN" = "1" ]; then
-    ShowMessage "  [MOD|sign] " `basename $NEWAPK`  " => " `basename "$APK"`
+    ShowMessage "  [MOD|sign] " `basename $NEWAPK`
     sign.sh $NEWAPK $APK &>> $LOG || exit 1
   else
-    ShowMessage "  [MOD|mv] " `basename $NEWAPK`  " => " `basename "$APK"`
+    ShowMessage "  [MOD|mv] " `basename $NEWAPK`
     mv $NEWAPK $APK &>> $LOG || exit 1
   fi
 
   cd - &> /dev/null
 fi
-
 exit 0
